moved some stuff around to prepare for writing the API
