+++ .gitignore
@@ -9,3 +9,4 @@ package-lock.json
 yarn.lock
 secrets/
 .vscode/
+*.log
