+++ src/services/temp/temp.js
@@ -1,22 +1,22 @@
-var InputData = require("../CostCenterDataController.js");
-var CreateUserData = require("../CreateAccountDataController");
+var InputData = require("./CostCenterDataController.js");
+var CreateUserData = require("./CreateAccountDataController");
 
 const newData = {
-     costCenter: "CC6526",
-     values: {
-         InputDate: "2018-01-01",
-         UnitsProduced: 2,
-         Defects: 2,
-         WorkerTotal: 2,
-         SInc_Num: 2,
-         QInc_Num: 2, 
-         SInc_Reason: "Lorem ipsum",
-         QInc_Reason: "Lorem ipsum",
-         HighUtil: "Lorem ipsum",
-         LoUtil: "Lorem ipsum",
-         Overtime: 2,
-         Downtime: 2,
-     }
+    costCenter: "CC6526",
+    values: {
+        InputDate: "2018-01-01",
+        UnitsProduced: 2,
+        Defects: 2,
+        WorkerTotal: 2,
+        SInc_Num: 2,
+        QInc_Num: 2,
+        SInc_Reason: "Lorem ipsum",
+        QInc_Reason: "Lorem ipsum",
+        HighUtil: "Lorem ipsum",
+        LoUtil: "Lorem ipsum",
+        Overtime: 2,
+        Downtime: 2,
+    }
 };
 
 const newUser = {
